@@ -1,13 +1,13 @@
 module github.com/appscodelabs/hugo-tools
 
-go 1.17
+go 1.18
 
 require (
-	github.com/appscode/static-assets v0.8.2-0.20220509092157-a2a01567a42c
+	github.com/appscode/static-assets v0.8.2-0.20221012112518-18b84d34c583
 	github.com/gohugoio/hugo v0.49.1
 	github.com/imdario/mergo v0.3.5
 	github.com/spf13/cast v1.3.1
-	github.com/spf13/cobra v1.1.3
+	github.com/spf13/cobra v1.4.0
 	github.com/spf13/pflag v1.0.5
 	golang.org/x/text v0.3.7
 	gomodules.xyz/go-sh v0.1.0
